Create README.md
Proven data analyst Intern with experience in mining, interpreting and presenting large datasets to
improve spend efficiency across organizations. Led the implementation of new data analysis
system and achieved a reduction in errors and increase in data accuracy. Adept in collaborating
with cross-functional teams and delivering projects on-time and on-budget.

Looking for Data Analyst job.